+++ eslint.config.js
@@ -1,0 +1,3 @@
+import baseConfig from "@arnaud-barre/eslint-config";
+
+export default baseConfig;